Add metrics storage class
